Refactor: Extracted different parsing types from the DialectInstanceParser into their own classes. Extracted JsonPointerResolution from DialectInstanceParser class to its own trait. Could be class.
